+++ tests/test_update_refpkg.py
@@ -14,12 +14,16 @@ class UpdaterTester(unittest.TestCase):
 
     def test_decide_length_filter(self):
         from treesapp import update_refpkg as ts_update_mod
-        min_length = ts_update_mod.decide_length_filter(self.ref_pkg, proposed_min_length=800)
-        self.assertEqual(800, min_length)
+        min_length = ts_update_mod.decide_length_filter(self.ref_pkg, proposed_min_length=500)
+        self.assertEqual(500, min_length)
 
         # Ensure the incorrectly-provided percentage is converted to a proportion
         min_length = ts_update_mod.decide_length_filter(self.ref_pkg, min_hmm_proportion=80)
         self.assertEqual(443, min_length)
+
+        # Ensure the minimum length provided is used, not the HMM proportion
+        min_length = ts_update_mod.decide_length_filter(self.ref_pkg, proposed_min_length=800, min_hmm_proportion=80)
+        self.assertEqual(800, min_length)
         return
 
     def test_simulate_entrez_records(self):